cscvs to tla changeset 43
Author: srivasta
Date: 2003/10/10 17:44:26
   * Fix a corner case when ucf is reinstalled after being removed, but not
purged, for a while, in this case all the data is out of
date. Rotating the dataset would have ucf asking questions all over,
but that is better than silently ignoring files of packages that were
purged while ucf was removed, and are being reinstalled. Correct
behaviour trumps the user being asked questions again. closes: Bug#215077

git-archimport-id: srivasta@debian.org--2003-primary/ucf--devel--0.1--patch-42