Emulate dpkg more closely when user opts not to install new configuration
When ucf replaces a configuration file, it always leaves the old
version behind as "conffile.dpkg-old", and then installed the
"distributed" file as "conffile", just *like dpkg* does. Now, also made
it so that when ucf asks the user a question, and they respond by
opting not to replace the configuration file, ucf creates
"conffile.dpkg-dist", as dpkg would have created.


git-archimport-id: srivasta@debian.org--2003-primary/ucf--devel--0.1--patch-46