A serious issue with command line options and debcoinf fixed
Debconf re-execs the calling script under the frontend. It passes in
$@ back to the script -- but this works only if the script has not
already mangled the arguments (say, by parsing the command line
destructuvely). Also, any unexported variables are lost and must be
re-calculated. The solution is to source the debconf confmodule before
the command line parsing. It really makes sense to stick it right at
the top, since anywork done before would need to be redone anyway.


git-archimport-id: srivasta@debian.org--2003-primary/ucf--devel--0.1--patch-64