Add support for displaying diffs using debconf itself
Users using a Graphical frontend are surprised when the display apparently
just blocks when they ask to see a diff (or 3-way diff) of the
configuration file being handled, when actually the diff is  displayed on
the terminal window ucf was run on. Until the debconf-escape utility and
the escape CAPB support, db_subst ran into newline and line length
issues. Patch from Michael Vogt.

Closes: Bug#325576

* ucf (show_diff): Added a function to show the diff using debconf

* ucf (newsum): Use it


git-archimport-id: srivasta@debian.org--lenny/ucf--devel--2.0--patch-2