There is a spurious global qualifier that make ucfq fail
Niko Tyni says ucfq /path/to/file never gives any output. It looks like
the problem is that ^/ is matched twice with the global option (m//g) on
the filename in ucf::process(), so the second one fails looking for the
next match.

I don't see the need for global match here, and removing the /g fixes
this for me.

* ucfq (process): Remove the /g qualifiers.

--This line, and those below, will be ignored--
Files to commit:
   ucfq

This list might be incomplete or outdated if editing the log
message was not invoked from an up-to-date changes buffer!

git-archimport-id: srivasta@debian.org--lenny/ucf--devel--3.0--patch-3