Major enhancements to functionality
Add a command line option to specify an alternate, caller provide debconf
template to use for prompting. Use db_metaget _template Choices-C to see
if the new template should be used.

Also, use db_x_loadtemplatefile to load our default templates, in case
the debconf DB had become corrupted.  This means we can stop using
debconf-loadtemplate, which was of a lower priority anyway.

In case of three way merge, exit with an error if the merge failed.

--This line, and those below, will be ignored--
Files to commit:
   ucfq
   ucf.1
   ucf
   ChangeLog

This list might be incomplete or outdated if editing the log
message was not invoked from an up-to-date changes buffer!

git-archimport-id: srivasta@debian.org--lenny/ucf--devel--3.0--patch-4