Sleep for 3 seconds after emitting warning
When there is no tty present (which means we cannot attach to the
keyboard), and there is no DISPLAY set, which means we cannopt fire off
another window, we cannot fork off a shell. This is not a bug, really,
but merely a reality. This patch sleeps for a couple of seconds so that
the error message we dump on STDERR can be read.

Closes: #543367

Signed-off-by: Manoj Srivastava <srivasta@debian.org>