Handle errors in three way merges
This fix is one that was first implemented in Ubuntu, and has only been
lightly modified. The problem was the handling of merge failures in three
way merges, ucf used to just pass the error through to the maintainer
script, and so a merge failure resulted in a maintainer script failure,
which could cascade in a upgrade. However, trapping the error, and
showing it via debconf, and then returning to the ucf debconf menu allows
users to resolve this issue as they please.

This patch implements that.

Signed-off-by: Steve Langasek <vorlon@debian.org>
Signed-off-by: Manoj Srivastava <srivasta@debian.org>