[upstream]: ucf fails when VERBOSE is set to 1 in /etc/ucf.conf
ucf fails with error code 20 when VERBOSE is set to 1 in
/etc/ucf.conf. Actually it is debconf's /usr/share/debconf/confmodule
which fails and subsequently topples ucf. The problem seems to be,
that some of the messages in ucf are written to stdout instead of
stderr, which confuses confmodule's function _db_cmd when reading
_db_internal_line. I believe the same is true when DEBUG is set > 0 in
ucf.conf, but I had no time to test it in more detail.

Fix thanks to Aljoscha Lautenbach <aljoscha.lautenbach@gmail.com>

Signed-off-by: Manoj Srivastava <srivasta@golden-gryphon.com>