feat(Helper): Add ucf helper functions
This commit introduces a new subdirectory ucf_helper_functions under
the examples directory. It contains a shell function include file
ucf_helper_functions.sh which introduces a quite advanced helper
function to handle lots of ucf conffiles. This code has been in use in
the aide package for some time and has not caused any problems.

The directory also includes a test suite which was used to develop the
code and can now be used to verify the code.

There is some some documentation in the 00README file in the examples.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>