bug(quote) Fix a quoting bug that triggered on dash
So this is something, one I didn't know about:
In bash and dash,

    var=$(something)

works as expected even if that ``something`` contains spaces, like in

    var=$(date '+%F %T')

Things are however different in dash if you're within a function and
want to limit the scope, i.e.

    local var=$(date '+%F %T')

Quite frankly, I haven't checked which shell implementation does not
follow the specification, if any, or if this is just another bashism.
Quoting fixes this

Signed-off-by: Manoj Srivastava <srivasta@debian.org>